Fixes for Unicode diagnostic messages for PostgreSQL & MySQL
Some drivers don't encode SQLSTATE and error messages with the same encoding.  So far it looks
like everyone might use UTF16LE.  This code handles UTF8, UTF16LE, and UTF16BE with no
configuration.  (It does not handle UTF32.)

The previous change converted the error message from bytes to Unicode, but missed the handling
of the 2nd error message.  When an attempt to concatenate was made, PyString_ConcatAndDel would
fail causing the HY0000 error.

+++ src/params.cpp
@@ -1530,7 +1530,7 @@ bool ExecuteMulti(Cursor* cur, PyObject* pSql, PyObject* paramArrayObj)
                 SQLGetStmtAttr(cur->hstmt, SQL_ATTR_APP_PARAM_DESC, &desc, 0, 0);
                 SQLSetDescField(desc, i + 1, SQL_DESC_TYPE, (SQLPOINTER)SQL_C_NUMERIC, 0);
                 SQLSetDescField(desc, i + 1, SQL_DESC_PRECISION, (SQLPOINTER)cur->paramInfos[i].ColumnSize, 0);
-                SQLSetDescField(desc, i + 1, SQL_DESC_SCALE, (SQLPOINTER)cur->paramInfos[i].DecimalDigits, 0);
+                SQLSetDescField(desc, i + 1, SQL_DESC_SCALE, (SQLPOINTER)(uintptr_t)cur->paramInfos[i].DecimalDigits, 0);
                 SQLSetDescField(desc, i + 1, SQL_DESC_DATA_PTR, bindptr, 0);
             }
             bindptr += cur->paramInfos[i].BufferLength + sizeof(SQLLEN);
